Fix sign compare warning

+++ pbar.h
@@ -101,7 +101,7 @@ inline void ProgressBar<It>::notify() {
   if (pos_old == pos)
     return;
   std::clog << left_delim_;
-  for (int i=0; i<width_; i++) {
+  for (size_t i=0; i<width_; i++) {
     if (i < pos)
       std::clog << symbol_;
     else if (i == pos)
